A clear README
The specific detail in the root level HTML’s is partly summarised
in the Announce, and the docs go into more detail.
Announce will serve as a proper README and we don’t clutter the
top level of the file hierarchy.

(We’re moving from a website/package to a package)

git rm *.html
git rm *.shtml
git mv Announce README